Add options menu entries for compatibility
I've been using the mod DeepRim which spawns maps that aren't considered player homes and I'd like to capture them.

I found the code responsible for skipping capturing them and added a check and an entry in the mod options that would allow those maps to be captured.

Edits are minimal.

Added option to use map name instead of world tile ID in filenames

This resolves the issue I had in being unable to use the mod to capture multiple maps on the same world tile (because before the names would conflict and cause "alt-#" to be added at random)